added constraint search with superimposed rectangle of occupied space

+++ src/utilities.hpp
@@ -48,7 +48,7 @@ struct Point
   }
 
   Point operator*(float scalar) const {
-    return {static_cast<int>(x * scalar), static_cast<int>(y * scalar)};
+    return {(x * scalar), (y * scalar)};
   }
 };
